+++ packages/server/lib/cloud/cy-prompt/CyPromptLifecycleManager.ts
@@ -7,9 +7,9 @@ import { isRetryableError } from '../network/is_retryable_error'
 import { asyncRetry } from '../../util/async_retry'
 import { postCyPromptSession } from '../api/cy-prompt/post_cy_prompt_session'
 import path from 'path'
-import os from 'os'
 import { readFile } from 'fs-extra'
 import { ensureCyPromptBundle } from './ensure_cy_prompt_bundle'
+import { parseHashFromBundleUrl } from '../bundles/parse_hash_from_bundle_url'
 import chokidar from 'chokidar'
 import { getCloudMetadata } from '../get_cloud_metadata'
 import type { CyPromptAuthenticatedUserShape, CyPromptServerOptions } from '@packages/types'
@@ -20,7 +20,7 @@ import type { ExitStepKey } from '../../util/graceful-exit'
 const debug = Debug('cypress:server:cy-prompt-lifecycle-manager')
 
 export class CyPromptLifecycleManager {
-  private static hashLoadingMap: Map<string, Promise<Record<string, string>>> = new Map()
+  private static hashLoadingMap: Map<string, Promise<{ manifest: Record<string, string>, cyPromptPath: string }>> = new Map()
   private static watcher: chokidar.FSWatcher | null = null
   private static teardown: ExitStepKey | null = null
 
@@ -186,9 +186,7 @@ export class CyPromptLifecycleManager {
     })
 
     if (!process.env.CYPRESS_LOCAL_CY_PROMPT_PATH) {
-      // The cy prompt hash is the last part of the cy prompt URL, after the last slash and before the extension
-      this.cyPromptHash = cyPromptSession.cyPromptUrl.split('/').pop()?.split('.')[0] as string
-      cyPromptPath = path.join(os.tmpdir(), 'cypress', 'cy-prompt', this.cyPromptHash)
+      this.cyPromptHash = parseHashFromBundleUrl(cyPromptSession.cyPromptUrl)
 
       let hashLoadingPromise = CyPromptLifecycleManager.hashLoadingMap.get(this.cyPromptHash)
 
@@ -196,13 +194,15 @@ export class CyPromptLifecycleManager {
         hashLoadingPromise = ensureCyPromptBundle({
           cyPromptUrl: cyPromptSession.cyPromptUrl,
           projectId,
-          cyPromptPath,
         })
 
         CyPromptLifecycleManager.hashLoadingMap.set(this.cyPromptHash, hashLoadingPromise)
       }
 
-      manifest = await hashLoadingPromise
+      const result = await hashLoadingPromise
+
+      manifest = result.manifest
+      cyPromptPath = result.cyPromptPath
     } else {
       cyPromptPath = process.env.CYPRESS_LOCAL_CY_PROMPT_PATH
       this.cyPromptHash = 'local'